+++ programming-1/problem-3/problem-3/comparator.h
@@ -16,10 +16,10 @@ typedef struct benchmark_result {
 } benchmark_res;
 
 double count_average(double *runtime, size_t num);
-double *run_sort(char sort, size_t calls_number, void *base, size_t num, size_t size, int(*compar)(void*, void*), void*(*prepare_data)(void*, int));
+double *run_sort(char sort, size_t calls_number, void *base, size_t num, size_t size, double(*compar)(void*, void*), void*(*prepare_data)(void*, int));
 double count_standard_dev_1(double *runtime, double average, size_t num);
 double count_standard_dev_2(double *runtime, double average, size_t num);
-benchmark_res benchmark(char sort, size_t calls_number, void *base, size_t num, size_t size, int(*compar)(void*, void*), void*(*prepare_data)(void*, int));
+benchmark_res benchmark(char sort, size_t calls_number, void *base, size_t num, size_t size, double(*compar)(void*, void*), void*(*prepare_data)(void*, int));
 void free_benchmark_result(benchmark_res result);
 void print_benchmark_result(FILE *file, benchmark_res result);
 size_t find_best_run(double *runtime, size_t num);
